requirements.txt
psycopg2-binary
python-dotenv
schedule
requests
boto3
bcrypt